Initial README
Added a readme that was originally changelog.as. I'll link people to the github rather than trying to have them dig in the game files for the changelog moving forward. The readme includes my own TODO list and some potential ideas for the future, even for some of my other mods, will be removed later