Top Barrier Changes
Made it so all blobs can go above the top barrier if it is not a cave map. Lightened the sky above max build limit a bit to allow for better vision while still being apparent where the top of the map is. Had to add special cases in ArcherLogic.as and ArcherAnim.as for grapple